+++ zedd-app/src/menuUtil.ts
@@ -1,7 +1,6 @@
 import { MenuItemConstructorOptions } from 'electron'
 import { AppState, Task } from './AppState'
 import { ClarityState } from './ClarityState'
-import { orderBy } from 'natural-orderby'
 
 export const suggestedTaskMenuItems = (
   state: AppState,
@@ -9,7 +8,7 @@ export const suggestedTaskMenuItems = (
   checked: Task,
   onClick: (t: Task) => void,
 ): MenuItemConstructorOptions[] => {
-  return orderBy(state.getSuggestedTasks(), (t) => t.name).map(
+  return state.getTasksForMenu().map(
     (t): MenuItemConstructorOptions => ({
       label: t.name,
       sublabel:
@@ -20,7 +19,10 @@ export const suggestedTaskMenuItems = (
         ),
       type: 'checkbox',
       checked: checked === t,
-      click: (x) => onClick(state.getTaskForName(x.label)),
+      click: (x) => {
+        onClick(state.getTaskForName(x.label))
+        state.notifyTaskInteraction(state.getTaskForName(x.label))
+      },
     }),
   )
 }